azure-function/requirements.txt
azure-functions
azure-storage-blob
requests
python-dateutil